requirements.txt
langchain == 0.1.17
langchain_google_genai 
fastapi 
uvicorn
gradio
python-dotenv